user_service/requirements.txt
Django==4.2
grpcio>=1.58.0
protobuf==5.29.2
psycopg2-binary>=2.9.9,<3.0 